+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add single hand device layouts
 - Code refactor with tangent-cc-lib
 
 ## [4.0](https://github.com/andy23512/monkeytype-cc-extension/releases/tag/4.0)