+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
+- Fix right space doesn't have key label and highlight bug
 - Add single hand device layouts
-- Code refactor with tangent-cc-lib
 - Export device layout
 
 ## [4.0](https://github.com/andy23512/monkeytype-cc-extension/releases/tag/4.0)